Improve components for Nextra
