+++ src/molecules/command-format.module.css
@@ -1,0 +1,5 @@
+.command {
+  font-size: 1.5rem;
+  font-height: 1.25;
+  margin: 2rem 0 1rem 0;
+}
