+++ src/molecules/command-format.tsx
@@ -1,0 +1,13 @@
+import style from './command-format.module.css';
+
+export interface CommandFormatProps {
+  children: string;
+}
+
+export const CommandFormat = ({
+  children
+}: CommandFormatProps): JSX.Element => (
+  <div className={style.command}>
+    <code>{children}</code>
+  </div>
+);
