+++ src/organisms/command-args.tsx
@@ -1,44 +1,14 @@
-import React, { ReactNode } from 'react';
+import React from 'react';
 
+import { CommandArg, CommandArgList } from '../molecules/command-arg-list';
+import { CommandFormat } from '../molecules/command-format';
 import { VersionBadge } from '../molecules/version-badge';
-
-export interface CommandArg {
-  name: string;
-  about: ReactNode;
-  versionAvailableFrom?: string;
-  defaultValue?: string;
-}
 
 export interface CommandArgsProps {
   versionAvailableFrom: string;
   commandName: string;
   args?: readonly CommandArg[];
 }
-
-const CommandArg = ({
-  name,
-  about,
-  versionAvailableFrom,
-  defaultValue
-}: CommandArg) => (
-  <>
-    <code>{name}</code>
-    {': '}
-    {about}
-    <ul>
-      {versionAvailableFrom && (
-        <li>
-          <VersionBadge>{versionAvailableFrom}</VersionBadge> から利用可能
-        </li>
-      )}
-      {defaultValue && (
-        <li>
-          デフォルト値: <code>{defaultValue}</code>
-        </li>
-      )}
-    </ul>
-  </>
-);
 
 export const buildCommandFormat = (
   commandName: string,
@@ -51,7 +21,7 @@ export const buildCommandFormat = (
       if (arg.defaultValue === undefined) {
         formatted += `<${arg.name}>`;
       } else {
-        formatted += `[${arg.name}=${arg.defaultValue}]`;
+        formatted += `[${arg.name}]`;
       }
     }
   }
@@ -65,17 +35,7 @@ export const CommandArgs = ({
 }: CommandArgsProps) => (
   <>
     <VersionBadge>{versionAvailableFrom}</VersionBadge> から利用可能
-    <h2>
-      <code>{buildCommandFormat(commandName, args)}</code>
-    </h2>
-    {args && (
-      <ul>
-        {args.map((arg) => (
-          <li key={arg.name}>
-            <CommandArg {...arg} />
-          </li>
-        ))}
-      </ul>
-    )}
+    <CommandFormat>{buildCommandFormat(commandName, args)}</CommandFormat>
+    {args && <CommandArgList args={args} />}
   </>
 );
